% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot.R
\name{data_preparation}
\alias{data_preparation}
\title{The built-in function of data preparation for mapping and plotting.}
\usage{
data_preparation(data, years, nutrient, counties = NULL, states = NULL,
  farm_type = NULL, input_type = NULL, lat_max = NULL, lat_min = NULL,
  long_max = NULL, long_min = NULL, FIPSs = NULL,
  overlap_state_county = TRUE, combine_state_county = FALSE, fun = NULL,
  facet = NULL, level = "county", na.rm = TRUE, ...)
}
\description{
The built-in function of data preparation for mapping and plotting.
}
\author{
Wenlong Liu
}
