% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot.R
\name{data_prepare_plot}
\alias{data_prepare_plot}
\title{The built-in function of data preparation for mapping and plotting.}
\usage{
data_prepare_plot(data = "us_fertilizer_county", fun = NULL, facet = NULL,
  level = "county", na.rm = TRUE, ...)
}
\description{
The built-in function of data preparation for mapping and plotting.
}
\author{
Wenlong Liu
}
