% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fertilizer}
\alias{fertilizer}
\title{fertilizer}
\format{A data frame with 31,279 rows and 12 variables:
\describe{
  \item{FIPS}{FIPS is a combination of state and county codes, in character format.}
  \item{State}{The state abbr. of U.S.}
  \item{County}{County name in U.S.}
  \item{ALAND}{The land area in each county, unit: km squared}
  \item{AWATER}{The water area in each county, unit: km squared}
  \item{INTPTLAT}{The latitude of centriod in each county, e.g. 32.53638}
  \item{INTPTLONG}{The longitude of centriod in each county, e.g. -86.64449}
  \item{Quantity}{The quantity of fertilizeation as N or P, e.g. kg N or kg P}
  \item{Year}{The year of estimated data, e.g. 1994}
  \item{Nutrient}{The fertilizer type, e.g. N or P}
  \item{Farm.Type}{The land use type of fertilizer, e.g. farm and nonfarm}
  \item{Input.Type}{The input type of nutrient, e.g. Fertilizer or Manure}
  ...
}}
\usage{
fertilizer
}
\description{
This is a sample dataset from usfertilizer.   Please visit <https://wenlong-liu.github.io/usfertilizer>. for more details.
}
\examples{
  require(ggfertilizer)
  data(fertilizer)
}
\keyword{datasets}
