% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(data, nutrient = NULL, input_type = NULL, farm_type = NULL,
  years = NULL, counties = NULL, states = NULL, lat_max = NULL,
  lat_min = NULL, long_max = NULL, long_min = NULL, FIPSs = NULL,
  overlap_state_county = TRUE, combine_state_county = FALSE)
}
\arguments{
\item{data}{the dataset to work with, generally the full usfertilizer or its subsets.}

\item{nutrient}{fertilizer type of interest, default: both Nitrogen and Phosphorus.}

\item{input_type}{the source of nutrient, e.g. fertilizer or manure.}

\item{farm_type}{where the fertilizer was applied. e.g. farm, nonfarm.}

\item{years}{start year to show data, e.g. 1994, 2000:2005.}

\item{counties}{counties of interest, defalut: all avaible counties.}

\item{states}{states of interest, defalt: all avaialble states.}

\item{lat_max}{the maximum latitude of area of interest.}

\item{lat_min}{the minmum latitude of area of interest.}

\item{long_max}{the maximum longitude of area of interest.}

\item{long_min}{the minimum longitude of area of interest.}

\item{FIPSs}{FIPS numbers of interest, defalut: all the counties.}

\item{overlap_state_county}{Logic. If true, the function will overlaping
the input of states and counties. If false, the function will return
results either in the states or in the counties.}

\item{combine_state_county}{Logic. If true, the county will be changed into
county, state, e.g. Wake, NC; If false, no changes.}
}
\value{
A tibble with tidy data.
}
\description{
This function will return the full dataset (or a subset based on input parameters) of county-level estimates
 of fertilizer application quantity.  The fertilizer application type includes farm and non-farms.
 The dataset was primarily cleaned into Tidy data.  Details are availalble via <https://cran.r-project.org/web/packages/usfertilizer/vignettes/Introduction.html>
}
\examples{
data = fertilizer
get_data(data, counties = "Wake")
get_data(data, years = 1990)
get_data(data, nutrient = "N", years = 2003, states = "NC", FIPSs = 37145)

}
\seealso{
\code{link(get_FIPS)}
}
\keyword{datasets}
\keyword{tidydata}
