% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_title.R
\name{get_title}
\alias{get_title}
\title{Automatically generate the title for the map or plot.}
\usage{
get_title(Year = NULL, Nutrient = NULL, Input_Type = NULL,
  Farm_Type = NULL, State = NULL, County = NULL, State_title = NULL,
  County_title = NULL, FIPS = NULL, level = "county")
}
\arguments{
\item{Year}{the temporal coverage of the data to visualize. See link for available period.}

\item{Nutrient}{the nutrient type of fertilizer from usfertilizer.  Has to be
N, P or both (NULL).}

\item{Input_Type}{the input source of fertilizer, should be manure, fertilizer.}

\item{Farm_Type}{the spatial source of fertilizer, should be farm, nonfarm or both.}

\item{State}{the states that will show in the plot,default: all states.}

\item{County}{the counties that will show in the plot, default: all counties.}

\item{State_title}{the user-defined title for state.}

\item{County_title}{the user-defined title for county.}

\item{FIPS}{FIPS numbers of interest, defalut: all the counties.}

\item{level}{the spatial resolution of map, should be county or state, default county.}
}
\value{
a ggtile obejct.
}
\description{
The title of map of plot heavily depends on the input parameters.
To help user generate the title, this function will guess the range and type of
inputs and provide a title.  Always, this title can be override by ggtitle().
}
\examples{
require(ggfertilizer)
Year = 2002
Nutrient = "N"
get_title(Year = Year, Nutrient = Nutrient)
}
\seealso{
\code{link(get_data), link(get_FIPS)}
}
