% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfertilizer.R
\docType{package}
\name{ggfertilizer}
\alias{ggfertilizer}
\alias{ggfertilizer-package}
\alias{ggfertilizer-package}
\title{ggfertilizer: Retrieve, Summarize and Visualize the Fertilizer Data in USA}
\description{
Summarize the count-level fertilizer data in USA from 1945 to 2012.  Users can quickly retrieve
fertilizer data in Tidy format. There are also built-in APIs for users to conduct spatial analysis and
provide reproducible research output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wenlong-liu/ggfertilizer}
  \item \url{https://wenlong-liu.github.io/ggfertilizer}
  \item Report bugs at \url{https://github.com/wenlong-liu/ggfertilizer/issues}
}

}
\author{
\strong{Maintainer}: Wenlong Liu \email{wliu14@ncsu.edu} (0000-0002-2271-3541)

}
\keyword{internal}
