% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fertilizer.R
\name{theme_map_fertilizer}
\alias{theme_map_fertilizer}
\title{Create base map theme for further use.
Adapted from theme maps.
https://www.rdocumentation.org/packages/ggthemes/versions/3.5.0/source}
\usage{
theme_map_fertilizer(base_size = 9, base_family = "")
}
\description{
Create base map theme for further use.
Adapted from theme maps.
https://www.rdocumentation.org/packages/ggthemes/versions/3.5.0/source
}
